Title: 특정 구간의 원소 평균값, Time: 167ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240515/특정 구간의 원소 평균값/elemental-mean-value-for-a-particular-interval.java
@@ -28,7 +28,7 @@ public class Main {
     public static boolean check(int s, int str, int las){
 
         
-        double avg = s / (las - str + 1.0);
+        double avg = s / (double)(las - str + 1);
         // if (s % (las - str +1) > 0){
         //     return false;
         // }
